safari-600.7-branch/LayoutTests/fast/loader/image-src-encoded.html

<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
             if (window.testRunner) {
                 testRunner.dumpAsText();
             }

             addEventListener("load", function() {
                 shouldBeTrue('document.getElementById("foo").clientWidth==25');
             }, false);
        </script>
    </head>

    <body id="body">
        <div>This test passes if the image below is displayed.</div>
        <img id="foo" src="resources/image%40test.png">
    </body>
</html>
