LayoutTests/fast/loader/image-src-encoded-expected.txt

PASS document.getElementById("foo").clientWidth==25 is true
This test passes if the image below is displayed.
